first check

+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -5,13 +5,13 @@ import java.util.Collection;
 
 public interface UserService {
 
-    UserDto getUserById(int id);
+    UserDto getUserById(Long id);
 
     Collection<UserDto> getAllUsers();
 
     UserDto createUser(UserDto user);
 
-    UserDto updateUser(int id, UserDto userDto);
+    UserDto updateUser(Long id, UserDto userDto);
 
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
 }
